outros cruds

+++ src/main/java/com/eventos/eventos_api/domain/service/CrudEventoService.java
@@ -1,14 +1,17 @@
 package com.eventos.eventos_api.domain.service;
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.eventos.eventos_api.domain.repository.EventoRepository;
+import com.eventos.eventos_api.domain.repository.ParticipanteRepository;
 import com.eventos.eventos_api.model.output.EventoOutput;
 
 import antlr.debug.Event;
 
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.transaction.Transactional;
@@ -16,7 +19,9 @@ import javax.transaction.Transactional;
 import lombok.AllArgsConstructor;
 
 import com.eventos.eventos_api.domain.exception.EntidadeNaoEncontradaException;
+import com.eventos.eventos_api.domain.exception.NegocioExeption;
 import com.eventos.eventos_api.domain.model.Evento;
+import com.eventos.eventos_api.domain.model.Participante;
 import com.eventos.eventos_api.domain.model.StatusEvento;
 
 @AllArgsConstructor
@@ -24,6 +29,7 @@ import com.eventos.eventos_api.domain.model.StatusEvento;
 public class CrudEventoService {
     
     private EventoRepository eventoRepository;
+    private ParticipanteRepository participanteRepository;
 
     public List<Evento> buscarTodosEventos(){
         return eventoRepository.findAll();
@@ -32,6 +38,18 @@ public class CrudEventoService {
     public Evento buscarUnicoEvento(Long eventoId){
         return eventoRepository.findById(eventoId)
         .orElseThrow(() -> new EntidadeNaoEncontradaException("Esse evento não existe!"));
+    }
+
+    //Loucura esse metodo aqui mas tenho que estudar ele
+    @Transactional
+    public List<Participante> buscarParticipantesDoEvento(Long eventoId){
+        List<Participante> participantesDoEvento = participanteRepository.findByEventoId(eventoId);
+
+        if (participantesDoEvento.isEmpty()) {
+            throw new NegocioExeption("Nenhum participante cadastrado nesse Evento!");
+        }
+
+        return participantesDoEvento;
     }
 
     @Transactional
@@ -53,6 +71,13 @@ public class CrudEventoService {
         if(!eventoRepository.existsById(eventoId)){
             throw new EntidadeNaoEncontradaException("Evento não encontrado!");
         }
+
+        long totalParticipantes = participanteRepository.countByEventoId(eventoId);
+
+        if (totalParticipantes > 0) {
+            throw new NegocioExeption("Não é possível excluir um evento que já possui participantes inscritos.");
+        }
+        
         eventoRepository.deleteById(eventoId);
     }
 }
